Feat: 채팅시 클라가 아닌 서버에서 petlevel 반환으로 설정

+++ src/main/java/com/loveforest/loveforest/domain/chat/dto/ChatMessageRequestDTO.java
@@ -21,5 +21,5 @@ public class ChatMessageRequestDTO {
     private Long coupleId; // 발신자 ID
 
     @Schema(description = "반려몽 레벨", example = "1~4")
-    private Long petLevel; // 반려몽 레벨
+    private int petLevel; // 반려몽 레벨
 }
